+++ src/app/views/dashboard-view/views/osc-automations-view/osc-automations-view.component.ts
@@ -1,10 +1,47 @@
-import { Component } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Subject, takeUntil } from 'rxjs';
+import {
+  AUTOMATION_CONFIGS_DEFAULT,
+  OscGeneralAutomationConfig,
+} from '../../../../models/automations';
+import { cloneDeep } from 'lodash';
+import { OpenVRService } from '../../../../services/openvr.service';
+import { AutomationConfigService } from '../../../../services/automation-config.service';
+import { OscScript } from '../../../../models/osc-script';
 
 @Component({
   selector: 'app-osc-automations-view',
   templateUrl: './osc-automations-view.component.html',
   styleUrls: ['./osc-automations-view.component.scss'],
 })
-export class OscAutomationsViewComponent {
-  activeTab: 'SLEEPING_ANIMATIONS' | 'GENERAL' = 'GENERAL';
+export class OscAutomationsViewComponent implements OnInit, OnDestroy {
+  private destroy$: Subject<void> = new Subject();
+  protected config: OscGeneralAutomationConfig = cloneDeep(AUTOMATION_CONFIGS_DEFAULT.OSC_GENERAL);
+
+  constructor(
+    private openvr: OpenVRService,
+    private automationConfigService: AutomationConfigService
+  ) {}
+
+  ngOnInit(): void {
+    this.automationConfigService.configs
+      .pipe(takeUntil(this.destroy$))
+      .subscribe(async (configs) => {
+        this.config = configs.OSC_GENERAL;
+      });
+  }
+
+  ngOnDestroy() {
+    this.destroy$.next();
+  }
+
+  async setOscScript(
+    event: 'onSleepModeEnable' | 'onSleepModeDisable',
+    script: OscScript | undefined
+  ) {
+    await this.automationConfigService.updateAutomationConfig<OscGeneralAutomationConfig>(
+      'OSC_GENERAL',
+      { [event]: script }
+    );
+  }
 }